Added Animations For Jumping
