+++ Assets/Animations/Player/PlayerFall.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9f498507fa58c9540b2e4b0460eed3ac
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
