+++ Assets/Animations/Player/PlayerJump.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5dd5f2f8b9de3e1449a737d67022f563
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
