+++ Assets/Animations/Player/PlayerJumpToFall.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1d3a991a5de4d3a469e63078bc476bf4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
