+++ Assets/Scripts/Player Scripts/PlayerAnimations.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 06352f2adfd99a544bb7d28c23d69e5b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
